requirements.txt
Django==2.0.5
djangorestframework==3.8.2
future==0.16.0
Pillow==5.1.0
pytz==2018.4
serializers==0.2.4
